refactor: remove redundant components, update ThemeToggle
