+++ src/components/ThemeToggle.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts">
+import { useSettingsStore } from "@/stores/settings"
+
+const settings = useSettingsStore()
+</script>
+
+<template>
+  <button @click="settings.toggleTheme">Theme: {{ settings.theme }}</button>
+</template>
+
+<style scoped></style>
