--- src/components/header/TheHeaderDocumentControl.vue
@@ -1,11 +1,0 @@
-<script>
-export default {
-  name: "TheHeaderDocumentControl",
-}
-</script>
-
-<template>
-  <div></div>
-</template>
-
-<style scoped></style>
