--- src/components/header/TheHeaderGeneralControl.vue
@@ -1,15 +1,0 @@
-<script lang="ts">
-import { defineComponent } from "vue"
-import ThemeToggle from "@/components/header/ThemeToggle.vue"
-
-export default defineComponent({
-  name: "TheHeaderGeneralControl",
-  components: { ThemeToggle },
-})
-</script>
-
-<template>
-  <div><ThemeToggle /></div>
-</template>
-
-<style scoped></style>
